+++ src/test/java/seedu/address/storage/JsonAddressBookStorageTest.java
@@ -2,7 +2,6 @@ package seedu.address.storage;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalPersons.ALICE;
 import static seedu.address.testutil.TypicalPersons.HOON;
@@ -19,13 +18,10 @@ import org.junit.jupiter.api.io.TempDir;
 import seedu.address.commons.exceptions.DataLoadingException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.person.Person;
-import seedu.address.model.tag.Tag;
 
 public class JsonAddressBookStorageTest {
     private static final Path TEST_DATA_FOLDER = Paths.get("src", "test", "data", "JsonAddressBookStorageTest");
     private static final String TEMP_FILE = "TempAddressBook.json";
-    private static final String TEST_FILE = "test.json";
 
     @TempDir
     public Path testFolder;
@@ -111,157 +107,5 @@ public class JsonAddressBookStorageTest {
     @Test
     public void saveAddressBook_nullFilePath_throwsNullPointerException() {
         assertThrows(NullPointerException.class, () -> saveAddressBook(new AddressBook(), null));
-    }
-
-    @Test
-    public void readAddressBookFromJsonString_validJson_success() throws Exception {
-        String validJson = "{"
-                                   + "\"persons\": ["
-                                   + "{"
-                                   + "\"name\": \"Alice Pauline\","
-                                   + "\"phone\": \"94351253\","
-                                   + "\"email\": \"alice@example.com\","
-                                   + "\"address\": \"123, Jurong West Ave 6, #08-111\","
-                                   + "\"tags\": [\"friends\"]"
-                                   + "}"
-                                   + "]"
-                                   + "}";
-
-        JsonAddressBookStorage storage = new JsonAddressBookStorage(testFolder.resolve(TEST_FILE));
-        ReadOnlyAddressBook addressBook = storage.readAddressBookFromJsonString(validJson);
-        Person currentPerson = addressBook.getPersonList().get(0);
-        assertEquals(1, addressBook.getPersonList().size());
-        assertEquals("Alice Pauline", currentPerson.getName().toString());
-        assertEquals("94351253", currentPerson.getPhone().toString());
-        assertEquals("alice@example.com", currentPerson.getEmail().toString());
-        assertEquals("123, Jurong West Ave 6, #08-111", currentPerson.getAddress().toString());
-        assertTrue(currentPerson.getTags().contains(new Tag("friends")));
-    }
-
-    @Test
-    public void readAddressBookFromJsonString_invalidJson_throwsDataLoadingException() {
-        String invalidJson = "{ not valid json }";
-        JsonAddressBookStorage storage = new JsonAddressBookStorage(testFolder.resolve(TEST_FILE));
-        assertThrows(DataLoadingException.class, () -> storage.readAddressBookFromJsonString(invalidJson));
-    }
-
-    @Test
-    public void readAddressBookFromJsonString_missingPersonFields_throwsDataLoadingException() {
-        String jsonWithMissingFields = "{"
-                                               + "\"persons\": ["
-                                               + "{"
-                                               + "\"name\": \"Alice Pauline\""
-                                               + "}"
-                                               + "]"
-                                               + "}";
-
-        JsonAddressBookStorage storage = new JsonAddressBookStorage(testFolder.resolve(TEST_FILE));
-        assertThrows(DataLoadingException.class, () -> storage.readAddressBookFromJsonString(jsonWithMissingFields));
-    }
-
-    @Test
-    public void readAddressBookFromJsonString_invalidPersonFormat_throwsDataLoadingException() {
-        String jsonWithInvalidPerson = "{"
-                                               + "\"persons\": ["
-                                               + "{"
-                                               + "\"name\": \"Alice'Pauline\","
-                                               + "\"phone\": \"invalid\","
-                                               + "\"email\": \"notanemail\","
-                                               + "\"address\": \"123 Street\","
-                                               + "\"tags\": []"
-                                               + "}"
-                                               + "]"
-                                               + "}";
-
-        JsonAddressBookStorage storage = new JsonAddressBookStorage(testFolder.resolve(TEST_FILE));
-        assertThrows(DataLoadingException.class, () -> storage.readAddressBookFromJsonString(jsonWithInvalidPerson));
-    }
-
-    @Test
-    public void readAddressBookFromJsonString_emptyJson_success() throws Exception {
-        String emptyJson = "{\"persons\": []}";
-        JsonAddressBookStorage storage = new JsonAddressBookStorage(testFolder.resolve(TEST_FILE));
-        ReadOnlyAddressBook addressBook = storage.readAddressBookFromJsonString(emptyJson);
-
-        assertEquals(0, addressBook.getPersonList().size());
-    }
-
-    @Test
-    public void readAddressBookFromJsonString_multiplePersons_success() throws Exception {
-        String multiplePersonsJson = "{"
-                                             + "\"persons\": ["
-                                             + "{"
-                                             + "\"name\": \"Alice Pauline\","
-                                             + "\"phone\": \"94351253\","
-                                             + "\"email\": \"alice@example.com\","
-                                             + "\"address\": \"123, Jurong West Ave 6, #08-111\","
-                                             + "\"tags\": [\"friends\"]"
-                                             + "},"
-                                             + "{"
-                                             + "\"name\": \"Hoon Meier\","
-                                             + "\"phone\": \"8482424\","
-                                             + "\"email\": \"stefan@example.com\","
-                                             + "\"address\": \"little india\","
-                                             + "\"tags\": []"
-                                             + "}"
-                                             + "]"
-                                             + "}";
-
-        JsonAddressBookStorage storage = new JsonAddressBookStorage(testFolder.resolve(TEST_FILE));
-        ReadOnlyAddressBook addressBook = storage.readAddressBookFromJsonString(multiplePersonsJson);
-
-        assertEquals(2, addressBook.getPersonList().size());
-        assertEquals("Alice Pauline", addressBook.getPersonList().get(0).getName().toString());
-        assertEquals("Hoon Meier", addressBook.getPersonList().get(1).getName().toString());
-    }
-
-    @Test
-    public void readAddressBookFromJsonString_emptyString_throwsDataLoadingException() {
-        JsonAddressBookStorage storage = new JsonAddressBookStorage(testFolder.resolve(TEST_FILE));
-        assertThrows(DataLoadingException.class, () -> storage.readAddressBookFromJsonString(""));
-    }
-
-    @Test
-    public void parse_validJson_success() throws Exception {
-        String validJson = "{"
-                                   + "\"persons\": ["
-                                   + "{"
-                                   + "\"name\": \"Alice Pauline\","
-                                   + "\"phone\": \"94351253\","
-                                   + "\"email\": \"alice@example.com\","
-                                   + "\"address\": \"123, Jurong West Ave 6, #08-111\","
-                                   + "\"tags\": [\"friends\"]"
-                                   + "}"
-                                   + "]"
-                                   + "}";
-
-        ReadOnlyAddressBook addressBook = JsonAddressBookStorage.parse(validJson);
-        Person currentPerson = addressBook.getPersonList().get(0);
-
-        assertEquals(1, addressBook.getPersonList().size());
-        assertEquals("Alice Pauline", currentPerson.getName().toString());
-        assertEquals("94351253", currentPerson.getPhone().toString());
-        assertEquals("alice@example.com", currentPerson.getEmail().toString());
-        assertEquals("123, Jurong West Ave 6, #08-111", currentPerson.getAddress().toString());
-        assertTrue(currentPerson.getTags().contains(new Tag("friends")));
-    }
-
-    @Test
-    public void parse_invalidJson_throwsDataLoadingException() {
-        String invalidJson = "invalid json";
-        assertThrows(DataLoadingException.class, () -> JsonAddressBookStorage.parse(invalidJson));
-    }
-
-    @Test
-    public void parse_emptyJson_success() throws Exception {
-        String emptyJson = "{\"persons\": []}";
-        ReadOnlyAddressBook addressBook = JsonAddressBookStorage.parse(emptyJson);
-
-        assertEquals(0, addressBook.getPersonList().size());
-    }
-
-    @Test
-    public void parse_null_throwsDataLoadingException() {
-        assertThrows(DataLoadingException.class, () -> JsonAddressBookStorage.parse(null));
     }
 }
